+++ game_patch/object/event.cpp
@@ -409,76 +409,19 @@ CodeInjection Event__turn_off_redirector_patch {
     }
 };
 
-void event_process_fixed_delay_patch(rf::GenericEvent* event) {
-
-    // fixed delay handling for events broken in stock game
-    if (af_rfl_version(rf::level.version) && event && event->delay_timestamp.elapsed()) {
-
-        if (event_debug_enabled) {
-            rf::console::print("Processing {} message in event {} ({}) (delayed)",
-                event->delayed_msg ? "ON" : "OFF", event->name, event->uid);
-        }
-
-        if (event->delayed_msg) {
-            switch (static_cast<rf::EventType>(event->event_type)) {
-                case rf::EventType::UnHide: {
-                    using EventTurnOnFn = void(__thiscall*)(rf::GenericEvent*);
-                    static EventTurnOnFn event_turn_on = reinterpret_cast<EventTurnOnFn>(0x004BCDD0);
-                    event_turn_on(event);
-                    break;
-                }
-                case rf::EventType::Make_Invulnerable: {
-                    using EventTurnOnFn = void(__thiscall*)(rf::GenericEvent*);
-                    static EventTurnOnFn event_turn_on = reinterpret_cast<EventTurnOnFn>(0x004BC7C0);
-                    event_turn_on(event);
-                    break;
-                }
-                default: {
-                    //xlog::warn("Something {} turned on", event->uid);
-                }
-            }
-        }
-        else {
-            switch (static_cast<rf::EventType>(event->event_type)) {
-                case rf::EventType::UnHide: {
-                    using EventTurnOffFn = void(__thiscall*)(rf::GenericEvent*);
-                    static EventTurnOffFn event_turn_off = reinterpret_cast<EventTurnOffFn>(0x004BCDE0);
-                    event_turn_off(event);
-                    break;
-                }
-                case rf::EventType::Make_Invulnerable: {
-                    using EventTurnOffFn = void(__thiscall*)(rf::GenericEvent*);
-                    static EventTurnOffFn event_turn_off = reinterpret_cast<EventTurnOffFn>(0x004BC880);
-                    event_turn_off(event);
-                    break;
-                }
-                default: {
-                    //xlog::warn("Something {} turned off", event->uid);
-                }
-            }
-        }
-
-        if (rf::event_type_forwards_messages(event->event_type)) {
-            using EventActivateLinksFn = void(__thiscall*)(rf::GenericEvent*, int, int, bool);
-            static EventActivateLinksFn event_activate_links = reinterpret_cast<EventActivateLinksFn>(0x004B8B00);
-            event_activate_links(event, event->trigger_handle, event->triggered_by_handle, event->delayed_msg);
-        }
-
-        event->delay_timestamp.invalidate();
-    }
-}
-
-CodeInjection EventUnhide__process_patch {
-    0x004BCDF0, [](auto& regs) {
-        rf::GenericEvent* event = regs.ecx;
-        event_process_fixed_delay_patch(event);
+CodeInjection EventUnhide__process_patch{
+    0x004BCDF0, 
+    [] (const auto& regs) {
+        rf::Event* const event = regs.ecx;
+        rf::Event__process(event);
     }
 };
 
-CodeInjection EventMakeInvulnerable__process_patch {
-    0x004BC8F0, [](auto& regs) {
-        rf::GenericEvent* event = regs.ecx;
-        event_process_fixed_delay_patch(event);
+CodeInjection EventMakeInvulnerable__process_patch{
+    0x004BC8F0,
+    [] (const auto& regs) {
+        rf::Event* const event = regs.ecx;
+        rf::Event__process(event);
     }
 };